These tests are for the database, not the app
